Simulation engine updated.
MaslClass now comes with single float type color components.
MaslSimulation is now a Swing window and its subclass, say
MaslSimulationSample, is self executable with a main() method.

+++ translation_samples/MaslClass.java
@@ -11,7 +11,7 @@ abstract public class MaslClass {
 	
 	public int x;
 	public int y;
-	public double r;
-	public double g;
-	public double b;
+	public float r;
+	public float g;
+	public float b;
 }
